Add optional exception
For some applications throwing an exception may be more convenient than
a false return value and printing a message.

This adds an optional :exception flag to the constructor in which case
a custom Exception will be thrown with the parsed error response if the
request is not successfull.

I've added Test::Mock as a dependency as that is more convenient for
testing stuff like this.