Send (and sign,) x-amz-security-token if provided.
If the client is using "temporary" credentials provided by the "Security Token Service"
( using, for example, [Kivuli](https://github.com/jonathanstowe/Kivuli) the security
token *must* be provided and be part of the signed headers, otherwise the client will
receive "InvalidAccessKeyId".

This is tested and found to work in an EC2 instance.

Fixes #8